cloud_storage_to_big_query_function/requirements.txt
# Function dependencies, for example:
# package>=version
pandas==1.3.5
google-cloud-storage==2.5.0
functions-framework==3.*
google-cloud-bigquery==3.3.5